Refactored https
